Fix compilation-find-file advice handling of directory
compilation-mode keeps track of the "current" directory in the output by
matching on messages such as make's "Entering directory ...", when this
happens the 'directory' variable is set, but the current code just ignores
it when checking whether we should search across project directories or
not. This means that in cases when the filename is actually unambigous
relative to the given directory we'll still go off and try to find it all
over the project which is less than ideal.

Instead of re-implementing the correct directory logic here we refactor
compilation-find-file-projectile-find-compilation-buffer to locally append
our project directories to the the global compilation-search-path and let
the real compilation-find-file fun do it's thing.

While this makes the behaviour more consistent with the original fun it
does mean that now we always have to get all the project dirs which could
be quite involved in large projects. Previously we could elide this in the
"exits" case.