Adds flag to fd to not print leading "./"
Fixes projectile-dir-files-alien on macOS

As I describe in #1783, the function `projectile-dir-files-alien` is not returning the same value in macOS as the function `projectile-dir-files-native`, because it is returning values with a leading "./". This PR fixes the `fd` command invocation adding the `--strip-cwd-prefix` to remove leading "./", and therefore fixing the issue reported in https://github.com/doomemacs/doomemacs/issues/6504.