[Fix #1708] Search hidden files with projectile-ripgrep
The behaviour of projectile-ripgrep was inconsistent depending on
which package was being used. If the rg package was being used, hidden
files were included in the search, while if the ripgrep package was
being used, hidden files were excluded. This commit ensures that
hidden files are now consistently included, which is consistent with
the behaviour of other searches (like grep).