Find project files bottom-up
Previously, the directory it found would depend on the order of
`projectile-project-root-files-bottom-up` (or the provided list).

given a directory structure

foo
├── .a
└── bar
    └── .a
    └── baz
        └── .b

`(projectile-root-bottom-up ".../foo/bar/baz/" '(".a" ".b"))` would return
`".../foo/bar/"`, because the current definition looks for any `".a"` before it
looks for ".b".

With this change, it traverses the directory structure once, looking for any
element of the list before moving up a level. So now the same call returns
`".../foo/bar/baz/"`, which matches the documented behavior of returning the
bottommost matched directory.

Fixes #1796