Update projectile-root-bottom-up tests
They should return the bottommost directory that contains a matching file, not
the bottommost directory for the first file matched.

Incidentally, reordered the files created for the test, so it’s easier to follow
the hierarchy. This also necessitated changing `projectile-test-with-files` so
that attempting to create a file before creating the directory that contains it
didn’t fail.