Add new custom variable 'projectile-cmd-hist-ignoredups'
Add new custom variable 'projectile-cmd-hist-ignoredups', which can be
used to tweak how duplicates are dealt with in projectile's command
history.  The custom variable is identical in behaviour to
'eshell-hist-ignoredups'.

Specifically, the existing default behavior is maintained with the
value of t, which means consecutive duplicates are ignored.  A value
of 'erase means only the last duplicate is kept, whilst a value of nil
means all duplicates are kept.